Extract brancher regex utility code from prowgen
